+++ modules/hocs/asSubmit.js
@@ -5,6 +5,6 @@ import { getContext } from 'recompose'
 export default function asSubmit(component: any): any {
   return getContext({
     submitForm: PropTypes.func,
-    isFormValid: PropTypes.bool
+    isFormValid: PropTypes.bool,
   })(component)
 }