+++ modules/index.js
@@ -1,4 +1,9 @@
+import Data from './components/Data'
+import Field from './components/Field'
 import Form from './components/Form'
+import Reset from './components/Reset'
+import State from './components/State'
+import Update from './components/Update'
 
 import asField from './hocs/asField'
 import asReset from './hocs/asReset'
@@ -13,7 +18,12 @@ import mapDataToValues from './helper/mapDataToValues'
 import formReducer from './model/reducers'
 
 export {
+  Data,
+  Field,
   Form,
+  Reset,
+  State,
+  Update,
   asField,
   asReset,
   asSubmit,
@@ -22,5 +32,5 @@ export {
   withState,
   formReducer,
   validateWithRequired,
-  mapDataToValues
+  mapDataToValues,
 }